+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 -   Deep cross networks: efficient ways of learning feature interactions.
+-   ScaNN integration: efficient approximate maximum inner product search for
+    fast retrieval.
 
 ## [0.2.0][2020-10-15]